created project structure and empty files
