+++ lib/features/dashboard/presentation/bloc/readable/readable_bloc.dart
@@ -1,0 +1,13 @@
+import 'package:bloc/bloc.dart';
+import 'package:equatable/equatable.dart';
+
+part 'readable_event.dart';
+part 'readable_state.dart';
+
+class ReadableBloc extends Bloc<ReadableEvent, ReadableState> {
+  ReadableBloc() : super(ReadableInitial()) {
+    on<ReadableEvent>((event, emit) {
+      // TODO: implement event handler
+    });
+  }
+}
