+++ lib/features/dashboard/presentation/bloc/readable/readable_event.dart
@@ -1,0 +1,5 @@
+part of 'readable_bloc.dart';
+
+sealed class ReadableEvent extends Equatable {
+  const ReadableEvent();
+}
