+++ lib/features/dashboard/presentation/bloc/readable/readable_state.dart
@@ -1,0 +1,10 @@
+part of 'readable_bloc.dart';
+
+sealed class ReadableState extends Equatable {
+  const ReadableState();
+}
+
+final class ReadableInitial extends ReadableState {
+  @override
+  List<Object> get props => [];
+}
