+++ lib/features/listing/presentation/bloc/description/description_event.dart
@@ -1,0 +1,5 @@
+part of 'description_bloc.dart';
+
+sealed class DescriptionEvent extends Equatable {
+  const DescriptionEvent();
+}
