+++ lib/features/listing/presentation/bloc/description/description_state.dart
@@ -1,0 +1,10 @@
+part of 'description_bloc.dart';
+
+sealed class DescriptionState extends Equatable {
+  const DescriptionState();
+}
+
+final class DescriptionInitial extends DescriptionState {
+  @override
+  List<Object> get props => [];
+}
