+++ lib/features/listing/presentation/bloc/listing/listing_event.dart
@@ -1,0 +1,5 @@
+part of 'listing_bloc.dart';
+
+sealed class ListingEvent extends Equatable {
+  const ListingEvent();
+}
