+++ lib/features/listing/presentation/bloc/listing/listing_state.dart
@@ -1,0 +1,10 @@
+part of 'listing_bloc.dart';
+
+sealed class ListingState extends Equatable {
+  const ListingState();
+}
+
+final class ListingInitial extends ListingState {
+  @override
+  List<Object> get props => [];
+}
